pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "google-cloud-translate>=3.20.2",
    "googletrans==4.0.0rc1",
    "langchain>=0.2.16",
    "langdetect>=1.0.9",
    "openai>=0.28.1",
    "pydub>=0.25.1",
    "pypdf2>=3.0.1",
    "reportlab>=4.3.1",
    "soundfile>=0.13.1",
    "streamlit>=1.44.1",
    "tts>=0.22.0",
    "whisper>=1.1.10",
]
